agent/requirements.txt
dedalus-labs>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
httpx>=0.25.0
